+++ src/i_sdlsound.cpp
@@ -1189,7 +1189,7 @@ static snddevice_t sound_sdl_devices[] = {
 
 sound_module_t sound_sdl_module = {
     sound_sdl_devices,
-    std::size(sound_sdl_devices),
+    static_cast<int>(std::size(sound_sdl_devices)),
     I_SDL_InitSound,
     I_SDL_ShutdownSound,
     I_SDL_GetSfxLumpNum,
