+++ src/strife/p_dialog.cpp
@@ -142,7 +142,7 @@ static dialogstateset_t dialogstatesets[] =
 };
 
 // Rogue stored this in a static global rather than making it a define...
-static int numdialogstatesets = std::size(dialogstatesets);
+static int numdialogstatesets = static_cast<int>(std::size(dialogstatesets));
 
 // Current dialog talker state
 static dialogstateset_t *dialogtalkerstates;
@@ -338,7 +338,7 @@ static rndmessage_t rndMessages[] =
 };
 
 // And again, this could have been a define, but was a variable.
-static int numrndmessages = std::size(rndMessages);
+static int numrndmessages = static_cast<int>(std::size(rndMessages));
 
 //=============================================================================
 //
